fixes #162
