+++ models/stat.js
@@ -1,0 +1,5 @@
+var bookshelf = require("../models/bookshelf");
+var Stat = bookshelf.Model.extend({
+	tableName: 'stats'
+});
+module.exports = Stat;
