+++ public/package.json
@@ -18,7 +18,7 @@
     "bootstrap": "^3.3.5",
     "can": "^2.3.15",
     "can-set": "^0.5.0-pre.2",
-    "can-connect": "^0.5.0-pre.1",
+    "can-connect": "^0.5.0-pre.2",
     "done-autorender": "^0.7.1",
     "done-component": "^0.4.0",
     "done-css": "^2.0.1",
