+++ services/teams.js
@@ -1,28 +1,7 @@
-
 var app = require("../services/app");
-var bookshelf = require("../models/bookshelf");
-var Player = require("./players");
+var Team = require("../models/team");
 var adminOnly = require( "./adminOnly" );
 
-var Team = bookshelf.Model.extend({
-	tableName: 'teams',
-	player1: function(){
-		return this.belongsTo(Player,"player1Id");
-	},
-	player2: function(){
-		return this.belongsTo(Player,"player2Id");
-	},
-	player3: function(){
-		return this.belongsTo(Player,"player3Id");
-	},
-	player4: function(){
-		return this.belongsTo(Player,"player4Id");
-	}
-});
-
-var Teams = bookshelf.Collection.extend({
-  model: Team
-});
 
 app.get('/services/teams', function(req, res){
 	Team.collection().query(req.query).fetch().then(function(teams){
@@ -55,5 +34,3 @@ app.post('/services/teams', adminOnly( "Must be an admin to create teams" ), fun
 		res.status(500).send(e);
 	});
 });
-
-module.exports = Team;
